requirements.txt
streamlit
langchain==0.2.1
langchain-openai==0.1.8
langchain-community==0.2.1
langchain_chroma
chromadb==0.5.0
pymupdf
azure-cognitiveservices-speech